# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Rent Room Information System API",
        "description": "API documentation for the Rent Room Information System",
        "contact": {
            "name": "Hinsy Uon",
            "url": "https://github.com/hinsyuon",
            "email": "it.hinsyuon@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/rooms": {
            "get": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Get list of rooms",
                "description": "Retrieve a paginated list of rooms with optional sorting and filtering.",
                "operationId": "bea681612754126f01807ff95647eb34",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search term to filter rooms by ID or room number.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter_col",
                        "in": "query",
                        "description": "Column to filter by [room_type_id, status, price_per_month].",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter_val",
                        "in": "query",
                        "description": "Value to filter the specified column by [room_type_id: 1 for Single, 2 for Double, 3 for Suite ], [status: 1 for Available, 2 for Booked]",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "price_min",
                        "in": "query",
                        "description": "Minimum price per month for filtering.",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "price_max",
                        "in": "query",
                        "description": "Maximum price per month for filtering.",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "sort_col",
                        "in": "query",
                        "description": "Column to sort by [id, room_number]",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "id"
                        }
                    },
                    {
                        "name": "sort_dir",
                        "in": "query",
                        "description": "Sort direction [asc or desc]",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "desc"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Current page number for pagination.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page for pagination.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get all rooms successfully"
                    },
                    "400": {
                        "description": "Bad Request - Validation errors"
                    }
                }
            }
        },
        "/api/room-types": {
            "get": {
                "tags": [
                    "Room Types"
                ],
                "summary": "Get all room types (Guest / Admin)",
                "operationId": "d688fcb7998e2add91986a9924200c0b",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search term to filter room types by ID, name, or description",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort_col",
                        "in": "query",
                        "description": "Column to sort [id, name]",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "id"
                        }
                    },
                    {
                        "name": "sort_dir",
                        "in": "query",
                        "description": "Sort direction [asc or desc]",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "desc"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Current page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get all room types successfully"
                    },
                    "400": {
                        "description": "Query Parameters failed validation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Room Types"
                ],
                "summary": "Create a new room type (Admin only)",
                "operationId": "0f40dd7add3fc834174fcc5911900a0f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Single"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "A room assigned to one person."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Room type created successfully"
                    },
                    "422": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/room-types/{id}": {
            "get": {
                "tags": [
                    "Room Types"
                ],
                "summary": "Get a room type by ID (Guest / Admin)",
                "operationId": "9535d689072e9302c71ecca20a918bc6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the room type to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get room type successfully"
                    },
                    "404": {
                        "description": "Room type not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Room Types"
                ],
                "summary": "Update a room type (Admin only)",
                "operationId": "1166d496a1a58a1ba342d9b297bbe6f8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the room type to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Single"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "A room assigned to one person."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Room type updated successfully"
                    },
                    "422": {
                        "description": "Invalid input"
                    },
                    "404": {
                        "description": "Room type not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Room Types"
                ],
                "summary": "Delete a room type",
                "operationId": "cc33e4c52544914924e4e5995ffff635",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the room type to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Room type deleted successfully"
                    },
                    "404": {
                        "description": "Room type not found"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Rooms",
            "description": "API Endpoints for Managing Rooms"
        },
        {
            "name": "Room Types",
            "description": "API Endpoints for Managing Room Types"
        }
    ],
    "components": {
        "securitySchemes": {
            "api_key_security_example": {
                "type": "apiKey",
                "description": "A short description for security scheme",
                "name": "api_key",
                "in": "header"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}